Add missing include.
On GCC > 4.6, ListenSocket will not compile:
"ListenSocket.h:285:17: error: there are no arguments to ‘close’
that depend on a template parameter, so a declaration of ‘close’
must be available [-fpermissive]"

This is caused by missing include unistd.h which imports 'close()'.
The actual close() function is masked with macro closesocket in
socket_include.h.